code/requirements.txt
openai
python-dotenv
pytest
pytest-cov
